requirements.txt
python-telegram-bot[webhooks]==20.8
requests==2.31.0
